requirements.txt
streamlit>=1.36.0
pandas>=2.2.2
openpyxl>=3.1.3
Pillow>=10.3.0
python-barcode>=0.15.1
arabic-reshaper>=3.0.0
python-bidi>=0.6.0
requests>=2.31.0
